+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[5.0.0-alpha.15](https://github.com/multiversx/mx-sdk-dapp/pull/1484)] - 2025-07-03
+
+- [Changed getDefaultNativeAuthConfig to accept an object as config](https://github.com/multiversx/mx-sdk-dapp/pull/1483)
+- [Added allowedProviders as list of strings](https://github.com/multiversx/mx-sdk-dapp/pull/1482)
+
 ## [[5.0.0-alpha.14](https://github.com/multiversx/mx-sdk-dapp/pull/1479)] - 2025-06-30
 
 - [Updated version of "mx-sdk-dapp-ui@0.0.8"](https://github.com/multiversx/mx-sdk-dapp/pull/1478)